requirements.txt
# Slack SDK
slack-bolt>=1.18.0
slack-sdk>=3.23.0

# Async support
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# HTTP client for Wren AI
httpx>=0.25.0

# Claude API (Anthropic)
anthropic>=0.39.0

# SQL parsing and validation
sqlparse>=0.4.4

# Configuration management
python-dotenv>=1.0.0

# Data manipulation and export
pandas>=2.0.0
numpy>=1.24.0

# Visualization (charts)
plotly>=5.18.0
kaleido>=0.2.1

# AWS Redshift connector (for direct queries if needed)
redshift-connector>=2.0.918

# Fuzzy string matching (for suggestions)
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0